+++ src/main/java/io/split/android/client/service/workmanager/splits/SplitsSyncWorkerParams.java
@@ -1,4 +1,4 @@
-package io.split.android.client.service.workmanager;
+package io.split.android.client.service.workmanager.splits;
 
 import androidx.work.WorkerParameters;
